Java code for Evaluating Postfix operation
Following is the algorithm for evaluation of postfix expressions:
Create a stack to store operands (or values).
Scan the given expression and do following for every scanned element.
If the element is a number, push it into the stack.
If the element is an operator, pop operands for the operator from the stack. Evaluate the operator and push the result back to the stack.
When the expression is ended, the number in the stack is the final answer.